requirements.txt
anthropic>=0.40.0
textual>=0.85.0
redis>=5.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0
click>=8.1.0
aiofiles>=23.0.0
